The Witness: Comply with new test base structure #5265
