+++ worlds/witness/test/test_door_shuffle.py
@@ -1,7 +1,7 @@
 from typing import cast
 
 from .. import WitnessWorld
-from ..test import WitnessMultiworldTestBase, WitnessTestBase
+from ..test.bases import WitnessMultiworldTestBase, WitnessTestBase
 
 
 class TestIndividualDoors(WitnessTestBase):
