+++ worlds/witness/test/test_ep_shuffle.py
@@ -1,4 +1,4 @@
-from ..test import WitnessTestBase
+from ..test.bases import WitnessTestBase
 
 
 class TestIndividualEPs(WitnessTestBase):
